refactor datetime helpers and add tests
